Move test classes to transport package
